refactor: Extract hierarchyToOptions
