+++ app/utils/hierarchy.ts
@@ -1,8 +1,10 @@
 import { ascending } from "d3";
 
+import { SelectTreeProps } from "@/components/select-tree";
 import { OptionGroup, Option } from "@/configurator";
 import { HierarchyParents } from "@/configurator/components/use-hierarchy-parents";
-import { HierarchyValue } from "@/graphql/query-hooks";
+import { HierarchyValue } from "@/graphql/resolver-types";
+import { sortTree } from "@/rdf/tree-utils";
 
 const asGroup = (
   parents: Omit<HierarchyValue, "depth" | "__typename" | "children">[]
@@ -28,4 +30,18 @@ export const makeOptionGroups = (
         ]
     )
     .sort((a, b) => ascending(a[1][0].depth, b[1][0].depth));
+};
+
+export const hierarchyToOptions = (hierarchy: HierarchyValue[]) => {
+  const transform = (h: HierarchyValue): SelectTreeProps["options"][number] => {
+    return {
+      ...h,
+      selectable: !!h.hasValue,
+      children:
+        h.children && h.children.length > 0
+          ? h.children.map(transform)
+          : undefined,
+    };
+  };
+  return sortTree(hierarchy).map((h) => transform(h));
 };
